# build/generated/sources/openapi/src/main/resources/openapi.yaml
openapi: 3.0.1
info:
  description: User API
  title: User
  version: "1"
servers:
- description: Local server
  url: http://localhost:8080/api
paths:
  /user:
    get:
      operationId: getAll
      responses:
        "200":
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/UserDTO'
                type: array
          description: OK
      summary: Get All Users
      x-accepts: application/json
    post:
      operationId: saveUser
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserDTO'
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserDTO'
          description: OK
      summary: Save User
      x-content-type: application/json
      x-accepts: application/json
  /user/{id}:
    get:
      operationId: getUserById
      parameters:
      - description: Unique Id of an User
        explode: false
        in: path
        name: id
        required: true
        schema:
          format: uuid
          type: string
        style: simple
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserDTO'
          description: OK
      summary: Get User By Id
      x-accepts: application/json
components:
  schemas:
    UserDTO:
      description: Data element for User
      example:
        firstName: firstName
        lastName: lastName
        id: 046b6c7f-0b8a-43b9-b35d-6489e6daee91
        email: email
      properties:
        id:
          format: uuid
          type: string
        email:
          type: string
        firstName:
          minLength: 3
          type: string
        lastName:
          type: string
      required:
      - email
      type: object
    Error:
      properties:
        code:
          format: int32
          type: integer
        message:
          type: string
      type: object